Revert "Split up Button and ButtonLink component"
This reverts commit fc67839318c1c3599319341087c656b9de2c1f0a.
