+++ packages/shared/components/Box.tsx
@@ -1,7 +1,6 @@
 import classNames from "classnames";
-import Button, { ButtonProps } from "./Button";
+import Button, { ButtonLinkProps, ButtonProps } from "./Button";
 import ButtonContainer from "./ButtonContainer";
-import ButtonLink, { ButtonLinkProps } from "./ButtonLink.tsx";
 import Heading, { type HeadingProps } from "./Heading";
 import RichText, { type RichTextProps } from "./RichText";
 
@@ -37,15 +36,9 @@ const Box = ({
       </div>
       {buttons && buttons.length > 0 && (
         <ButtonContainer>
-          {buttons.map((button) => {
-            if ("href" in button) {
-              return <ButtonLink key={button.text} {...button} />;
-            } else if ("onClickCallback" in button) {
-              return <Button key={button.text} {...button} />;
-            } else {
-              return null;
-            }
-          })}
+          {buttons.map((button) => (
+            <Button key={button.text ?? button.href} {...button} />
+          ))}
         </ButtonContainer>
       )}
     </div>
